"Type ':quit' to quit topydo."
Currently in the Columns UI, when you hit Control-C, the message displayed is "Use the 'quit' command to exit topydo". This led me to fruitlessly hit "q" and "quit" a bunch of times. I propose changing the message to "Type ':quit' to quit topydo" to make it more obviously exactly how to quit.